(3/n scheduler-refactor) Remove dependence on single repository in ScheduleStorage
Summary:
We previously used to store `repository_name` as a private variable on the `FilesystemScheduleStorage` class.

This diff changes it so that the repository name is not stored as a private variable, but instead passed to each of the functions directly. This makes it easier to move `SchedulerStorage` to the instance since we don't have access to the repository name when we initialize `ScheduleStorage` on the instance.

The majority of this diff is changing references to `self._repository_name` to be the argument `repository_name`.

Test Plan: unit

Reviewers: alangenfeld, prha

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D1890