Re-order schedule rotues in dagit
Summary:
In the Dagit top-level router, the `/schedules` route was being defined before the `/schedules/:scheduleName` route, so routes that should have matched the latter were being matched by the former.

This diff reorders the route so they are matched in the correct order.

Test Plan: Load dagit and click on schedule

Reviewers: #ft, themissinghlink

Reviewed By: #ft, themissinghlink

Subscribers: themissinghlink

Differential Revision: https://dagster.phacility.com/D1908