eliminate gbfs code
Summary:
This PR eliminates the gbfs code because it turns out the real time feed does not connect with the trip data at all. However, we have paths forward to demo the bay-bikes data with daily, monthly, and scheduled retraining pipelines so we will be able to effectively test out the partitions and backfill systems. This also addresses https://github.com/dagster-io/dagster/issues/2029.

All this code is saved in prior commits so we can get it back later if need be.

Test Plan: unit

Reviewers: schrockn, max

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D1922