skip loading resources when not available on the wider pipeline context
Summary:
We currently only load resources on the pipeline context that are required by the step solids that are about to execute.  However, for resume/retry, we perform copy intermediate operations for step contexts that provide inputs for solids that we are about to execute.  This means that we will construct step contexts (in order to yield the copy event) that spin up resources even though they will not be used.  This is because solid-specific resources are not used for the copy operation, only the system storage resources.

This diff ignores unknown resources while constructing the step context, in order to bypass this issue.  We already do load time checks that all the sufficient resources are configured on every mode.

Test Plan: Added pytest case for retrying resource dependent solids

Reviewers: alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D1979