Ensure zoom-to-fit scale is always >0 [2152]
Summary:
The error shown in https://github.com/dagster-io/dagster/issues/2152 is caused by the Gaant visualization computing a zoom-to-fit size of <=0 and then attempting to transform it logarithmically based on your preferred zoom [1-100] (producing NaN as shown in the screnshot). Math.log(0) and Math.log(-anything) are both invalid and we just need to make sure that both the elapsed time and the width of the view are positive here.

I broke this onto multiple lines and added local vars to name the divisor and divided and make this easier to follow.

Test Plan: run tests

Reviewers: prha, schrockn, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D2030