Refactor generate_training_set pipeline.
Summary:
This refactor adds a generate_training_set pipeline which downloads the weather and trip tables and produces a training set which is then uploaded to a gcs bucket. This PR cleans up a ton of unnused code and adds a pipeline test.

Call Out: Builds are breaking because of the gcp stuff.

Call Out 2: This revision brings in the `LocalClient` resource in favor of the old resources. The reason for this is that we want to actually mirror **persistent** cloud storage on our local file system. We need this because in a demo, we can't expect folks to have a gcs bucket provisioned. However, we also cannot use temporary files or temporary directories because we need these files to persist across pipeline runs (AKA the `generate_training_set` pipeline uploads a pickled training set to a directory, and the `train_model` pipeline -next revision- downloads that pickled training set).

Test Plan: pipeline integration test.

Reviewers: schrockn, alangenfeld, nate

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D2071