GraphQL PipelineSnapshotMixin
Summary:
This provides both code-sharing and for a common
interface for active pipelines and pipeline snapshots.

It creates:

	1)  A new type PipelineSnapshot,
	2)  A new graphql interface, shared between IPipelineSnapshot
and DauphinPipeline.
	3) A new mixin, DauphinPipelineSnapshotMixin to share
	implementation of IPipelineSnapshot

Graphene has some strange properties that make it so that you cannot
implement ABCs nor use properties in an overridable way. So the way
the mixin works is that the target classes have to have a method
get_pipeline_index()

Test Plan: BK

Reviewers: alangenfeld, themissinghlink

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D2355