Replace the playground's execution plan preview with a more useful config map
Summary:
The usefulness of the playground's run preview has been pretty limited so far and this diff replaces it with a new "summary" panel that shows you which parts of your config are valid, which are missing, and which contain errors. You can click errors to jump to them, and click sections to jump to them (though the top panel always shows the entire document - sub-"views" might be an interesting stacked diff). The "Solids" section of the summary is not currently a representation of the running order or anything like that - I think it'd be interesting, but I'm not totally convinced that the whitespace required to render a plan / gaant diagram, etc. is worth it in this view.

I also made the errors a more straightforward list so we don't run into issues with them being clipped off the screen.

{F107933}

Test Plan: Updated snapshots

Reviewers: prha, alangenfeld, schrockn

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D2368