Only allow raw config values as default_value
Summary:
This changes the semantics of the default value of a field
in config to only allow the pre-processed raw value rather than the
post-processed value. This is only pertinent to enums in today's system,
and we have a special error message to make that clear.

This will allow default value to be readily hashable, as well as
to potentially show previews of default values in dagit and other tools.

Depends on D2387

Test Plan: BK

Reviewers: alangenfeld, max

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D2388