#2352 Fix incorrect step state after re-execution
Summary:
issue: https://github.com/dagster-io/dagster/issues/2352
we will drop the step from metadata.steps if it's a ObjectStoreOperationEvent and op='CP_OBJECT'

The initial state of every step is "preparing". The re-execution needed to copy the intermediate object of "read_csv", which also fired a `ObjectStoreOperationEvent` log. So the skipped step read_csv appeared in the log with the initial state.
{F115401}

Test Plan:
- screenshot
{F116199}
- bk

Reviewers: schrockn, prha, bengotow

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D2472