(add-snapshot-persistence-3) Adding db-based snapshot persistence
Summary:
This adds a snapshot table to run storage. This sets up
the ability to persist pipeline and execution plan snapshots
alongside run, to allow for auditing, historical viewing, caching
representations of pipelines from containers, and so forth.

A few things happening in here:

1) The actual database code.
2) As part of that, started a pattern of being paranoid of what is in the database and logging whenever something doesn't conform. E.g. see defensively_unpack_pipeline_snapshot_query and defensively_get_single_row
3) Alembic migrations and snapshots for postgres and sqlite

Next stage of this will be actually changing the create run codepaths
within DagsterInstance to actually persist the snapshots in production
codepaths.

Depends on D2524

Test Plan: BK

Reviewers: prha, max

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D2496