(add-snapshot-persistence-4) Persisting snapshot in startPipelineExecution and execute_pipeline
Summary:
This completes the step of actually persisting pipeline
snapshots along with runs and having them be fetchable.

Biggest change here is that this adds a new class PipelineRunCreationData
that will serve as the public API for run creation, wholly separate from
PipelineRun, which is the pure data object that is serialized into the
data.

This is critical because it allows the instance to handle the
orchestration of pipeline creation and then it is responsible
for creating the PipelineRun.

In general, we were always going to need a separate API for pipeline
run creation the moment that we introduced any sort of normalization
or a computed field.

Depends on D2496

Test Plan: BK

Reviewers: max, prha, sashank

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D2514