(add-snapshot-persistence-5) GraphQL Query persisted snapshots
Summary:
This diff prepares the GraphQL layer to properly back
the snapshot queries. A few changes here:

- Introduces pipelineSnapshotOrError
- Introduces PipelineSnapshotNotFound error for above
- Changes implementation of pipelineSnapshot (and new
pipelineSnapshotOrError) to operator on snapshot id's rather
than pipeline name (which was just there for testing)
- Move runs to be in Pipeline only for now since its implementation
operators on pipeline names only. (This may be re-introduced for
snapshots to query by snapshot_id when that is needed)

Depends on D2514

Test Plan: BK

Reviewers: themissinghlink, prha, max

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D2531