(add-snapshot-persistence-6) Refactor pipeline snapshot tests as a fixture
Summary:
This mimics the structure of TestRunStorage, which I find quite
pleasant to work with.

I believe we should restructure nearly all of our graphql tests to use
this, and also introduce a battery of tests against the postgres implementation
as well.

This will serve as a really nice, leveraged layer of integration tests for the
entire product, assumign our graphql tests keep up-to-date.

Depends on D2531

Test Plan: BK

Reviewers: themissinghlink, sashank, max

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D2535