Make PartitionSet / Preset a part of session, improve selection mechanism
Summary:
This diff updates the Preset and Partition Set pickers in the playground to make it more clear that presets and partition sets impact the subset and mode settings, and arranges them in a better order. There's a high level diagram of the design here: https://excalidraw.com/#room=be7f2c1ce574eb6db535,25E6rMooMGIa_ZQhRAdb8w

- The preset / partition set dropdown now shows which mode / subset each one defines, so it's clear that when you choose an item it will overwrite those settings in your session.

- Changing the preset or partition no longer creates a new tab every time. I think this is fine because it's new position in the top left means it should be the first thing you do when creating a new tab. We might want to provide a "Duplicate Tab" behavior if folks want the old behavior?

- The partition picker is now a typeahead, which I think is slightly nicer for long sets of date options, etc.

- The partition picker has a more sophisticated loading state, because some partition sets like `generate_training_set_and_train_model_daily` take a while to load.

This is subtle, but the preset and partition set are now part of the session's stored state. This means that if you pick a preset and reload the page, the preset picker is still populated (rather than being a populate-and-forget action). This seems nice but at some point, if you modify your config enough, you're not really working from the preset that is displayed anymore. At what point if any should the Preset: picker be cleared if it's no longer representative? This might not actually happen.

{F117506}
{F117507}
{F117508}

Test Plan: Updated snapshots

Reviewers: prha, alangenfeld, schrockn

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D2536